pyproject.toml
[project]
name = "langchain-chain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-google-vertexai>=3.2.0",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "ruff>=0.14.10",
]
